+++ commons/util/src/test/java/de/learnlib/util/ExperimentTest.java
@@ -23,7 +23,6 @@ import de.learnlib.oracle.EquivalenceOracle.DFAEquivalenceOracle;
 import de.learnlib.query.DefaultQuery;
 import de.learnlib.statistic.Statistics;
 import de.learnlib.statistic.StatisticsService;
-import de.learnlib.util.Experiment.DFAExperiment;
 import net.automatalib.alphabet.Alphabet;
 import net.automatalib.alphabet.impl.Alphabets;
 import net.automatalib.automaton.fsa.DFA;